# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: College Resource Sharing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List resources with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: search
          in: query
          schema:
            type: string
        - name: collegeId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: branch
          in: query
          schema:
            type: string
        - name: semester
          in: query
          schema:
            type: integer
        - name: fileType
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, popular, mostDownloaded, topRated]
            default: newest
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourcesListResponse"

    post:
      operationId: createResource
      tags: [resources]
      summary: Upload a new resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceRequest"
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /resources/{id}:
    get:
      operationId: getResource
      tags: [resources]
      summary: Get a resource by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateResource
      tags: [resources]
      summary: Update a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResourceRequest"
      responses:
        "200":
          description: Resource updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"

    delete:
      operationId: deleteResource
      tags: [resources]
      summary: Delete a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /resources/{id}/download:
    post:
      operationId: incrementDownload
      tags: [resources]
      summary: Increment download count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Download count incremented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /resources/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: Get comments for a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /users/{id}/resources:
    get:
      operationId: getUserResources
      tags: [users]
      summary: Get resources uploaded by user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
...
  /admin/resources:
    get:
      operationId: adminListResources
      tags: [admin]
      summary: List all resources (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourcesListResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Resource:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        fileUrl:
          type: string
        subject:
          type: string
          nullable: true
        branch:
          type: string
          nullable: true
        semester:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        downloadCount:
          type: integer
        viewCount:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
        uploaderId:
          type: integer
        uploaderName:
          type: string
        collegeId:
          type: integer
          nullable: true
        collegeName:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        averageRating:
          type: number
          nullable: true
        ratingCount:
          type: integer
        isBookmarked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, fileName, fileType, fileSize, fileUrl, downloadCount, viewCount, status, uploaderId, uploaderName, ratingCount, isBookmarked, createdAt, tags]
    ResourceDetail:
      allOf:
        - $ref: "#/components/schemas/Resource"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
            userRating:
              type: integer
              nullable: true
          required: [comments]
    ResourcesListResponse:
      type: object
      properties:
        resources:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [resources, total, page, totalPages]
    CreateResourceRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        fileUrl:
          type: string
        subject:
          type: string
        branch:
          type: string
        semester:
          type: integer
        tags:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        collegeId:
          type: integer
      required: [title, fileName, fileType, fileSize, fileUrl]
    UpdateResourceRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        branch:
          type: string
        semester:
          type: integer
        tags:
          type: array
          items:
            type: string
        categoryId:
          type: integer
    Comment:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        userId:
          type: integer
        userName:
          type: string
        resourceId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, content, userId, userName, resourceId, createdAt]
    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
      required: [content]
...