# lib/api-zod/src/generated/types/resourceDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Resource Sharing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Comment } from "./comment";
import type { Resource } from "./resource";

export type ResourceDetail = Resource & {
  comments: Comment[];
  userRating?: number | null;
};