# lib/api-zod/src/generated/types/updateResourceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Resource Sharing Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateResourceRequest {
  title?: string;
  description?: string;
  subject?: string;
  branch?: string;
  semester?: number;
  tags?: string[];
  categoryId?: number;
}